refactor: extract StringInterner to proximadb-interner + move metadata_types to data-model
Completes the metadata_types shared-dep unlock (third increment of the
hermetic-leaf unlock in the root-crate decomposition track; follows #1071 and
#1075).

1. Extract StringInterner from src/storage/cache/orchestrator.rs into a new
   foundation crate `proximadb-interner` (XxHash64-keyed DashMap dedup of
   Arc<str>). orchestrator.rs re-exports it so the
   `crate::storage::cache::orchestrator::StringInterner` path resolves unchanged.
   StringInterner is hermetic (deps: dashmap + proximadb-kernel's XxHash64).

2. Move metadata_types.rs (MetadataValue / TypedMetadata) into the
   `proximadb-data-model` foundation crate (root keeps a pure `pub use` shim).
   This was previously blocked (#1075) because metadata_types:162 referenced
   `crate::storage::cache::orchestrator::StringInterner`; now that StringInterner
   is a leaf, metadata_types depends on `proximadb_interner` + `proximadb_proto`
   (the 9 `crate::proto` refs also retargeted to the proto crate).

metadata_types is a shared root-internal dep of the ORION graph engine and the
pgwire layer; moving it to a leaf removes another recurring dependency that
would force those cohorts into a cyclic root dependency when extracted next.

Pure relocation + re-export, no behavior change. Now-unused imports
(XxHash64, top-level std::hash::{...}, top-level std::sync::atomic::{...})
removed from orchestrator.rs.
