+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "crates/foundation/proximadb-clustering-kernel",
     "crates/foundation/proximadb-memory-pool",
     "crates/foundation/proximadb-kernel",
+    "crates/foundation/proximadb-interner",
     "crates/foundation/proximadb-bloom",
     "crates/foundation/proximadb-pipeline-operator",
     "crates/foundation/proximadb-proto",
@@ -179,6 +180,7 @@ proximadb-distance-kernel = { path = "crates/foundation/proximadb-distance-kerne
 proximadb-clustering-kernel = { path = "crates/foundation/proximadb-clustering-kernel" }
 proximadb-memory-pool = { path = "crates/foundation/proximadb-memory-pool" }
 proximadb-kernel = { path = "crates/foundation/proximadb-kernel" }
+proximadb-interner = { path = "crates/foundation/proximadb-interner" }
 proximadb-bloom = { path = "crates/foundation/proximadb-bloom" }
 proximadb-search-types = { path = "crates/foundation/proximadb-search-types" }
 proximadb-tenant = { path = "crates/foundation/proximadb-tenant" }
@@ -549,6 +551,7 @@ toml = "0.8"
 anyhow = "1.0"
 thiserror = "1.0"
 proximadb-kernel = { path = "crates/foundation/proximadb-kernel" }
+proximadb-interner = { path = "crates/foundation/proximadb-interner" }
 proximadb-bloom = { path = "crates/foundation/proximadb-bloom" }
 proximadb-search-types = { path = "crates/foundation/proximadb-search-types" }
 proximadb-tenant = { path = "crates/foundation/proximadb-tenant" }
