+++ crates/foundation/proximadb-interner/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "proximadb-interner"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB foundation-tier string interner (StringInterner) — XxHash64-keyed DashMap dedup of Arc<str> shared across metadata/cache layers"
+
+[lib]
+name = "proximadb_interner"
+path = "src/lib.rs"
+
+[dependencies]
+dashmap.workspace = true
+proximadb-kernel.workspace = true
